requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
numpy==1.20.0
pandas==1.2.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.3
geopandas
rasterio==1.2.10
fpdf==1.7.2
flask==2.0.3
flask_sqlalchemy==2.5.1